requirements.txt
absl-py==1.3.0
attrs==23.1.0
backcall==0.2.0
blessed==1.20.0
cachetools==5.2.0
certifi 
charset-normalizer==3.1.0
click==8.1.3
cycler==0.11.0
decorator==5.1.1
docker-pycreds==0.4.0
einops==0.6.1
ffmpeg-python==0.2.0
filelock==3.9.0
Flask-Login==0.6.2
Flask-WTF==1.0.1
fonttools==4.38.0
ftfy==6.1.1
future==0.18.3
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.15.0
google-auth-oauthlib==0.4.6
gpustat==1.1
grpcio==1.51.1
h5py==3.8.0
idna==3.4
importlib-metadata==6.0.0
ipython==7.34.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonlines==3.1.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
nncore==0.3.6
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-ml-py==11.525.131
oauthlib==3.2.2
opencv-python==4.8.0.76
packaging==22.0
pandas==1.3.5
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
promise==2.3
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
regex==2022.10.31
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml.clib==0.2.7
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
sentry-sdk==1.12.1
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
sklearn==0.0.post5
smmap==5.0.0
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tokenizers==0.13.2
torchtext==0.10.0
tqdm==4.65.0
traitlets==5.8.0
typing_extensions==4.4.0
urllib3==2.0.3
wcwidth==0.2.5
Werkzeug==2.2.2
wordcloud==1.8.0
WTForms==3.0.1
zipp==3.11.0
